🎨 move registering/deregistering instance to ghost start/stop commands
- ensures that instance is deregistered when it is stopped
