+++ lib/commands/run.js
@@ -29,7 +29,7 @@ module.exports.execute = function execute() {
             return knexMigrator.migrate();
         }
     }).then(() => {
-        instance = new Instance(this.ui, config, processManager);
+        instance = new Instance(this.ui, processManager);
     }).catch((error) => {
         processManager.error(error.message);
     });
