+++ lib/commands/start.js
@@ -3,6 +3,19 @@ const Config = require('../utils/config');
 const checkValidInstall     = require('../utils/check-valid-install');
 const resolveProcessManager = require('../utils/resolve-process');
 
+function register(config, environment) {
+    let systemConfig = Config.load('system');
+    let instances = systemConfig.get('instances', {});
+
+    instances[process.cwd()] = {
+        mode: environment,
+        url: config.get('url'),
+        port: config.get('server.port'),
+        process: config.get('process')
+    };
+    systemConfig.set('instances', instances).save();
+}
+
 module.exports.execute = function execute(options) {
     checkValidInstall('start');
 
@@ -17,6 +30,7 @@ module.exports.execute = function execute(options) {
 
     let processManager = resolveProcessManager(config);
     let start = Promise.resolve(processManager.start(process.cwd(), this.environment)).then(() => {
+        register(config, this.environment);
         cliConfig.set('running', this.environment).save();
         // TODO: add process info to global cli file so Ghost-CLI is aware of this instance
         return Promise.resolve();
