+++ lib/commands/stop.js
@@ -3,6 +3,13 @@ const Config = require('../utils/config');
 const checkValidInstall     = require('../utils/check-valid-install');
 const resolveProcessManager = require('../utils/resolve-process');
 
+function deregister() {
+    let systemConfig = Config.load('system');
+    let instances = systemConfig.get('instances', {});
+    delete instances[process.cwd()];
+    systemConfig.set('instances', instances).save();
+}
+
 module.exports.execute = function execute(options) {
     checkValidInstall('stop');
 
@@ -17,6 +24,7 @@ module.exports.execute = function execute(options) {
     let config = Config.load(`config.${cliConfig.get('running')}.json`);
     let processManager = resolveProcessManager(config);
     let stop = Promise.resolve(processManager.stop(process.cwd())).then(() => {
+        deregister();
         cliConfig.set('running', null).save();
         return Promise.resolve();
     });
