+++ lib/utils/Instance.js
@@ -1,11 +1,9 @@
 'use strict';
 const spawn = require('child_process').spawn;
-const Config = require('./config');
 
 class Instance {
-    constructor(ui, config, processManager) {
+    constructor(ui, processManager) {
         this.ui = ui;
-        this.config = config;
         this.processManager = processManager;
 
         this.child = spawn(process.execPath, ['current/index.js'], {
@@ -20,7 +18,6 @@ class Instance {
 
         this.child.on('message', (message) => {
             if (message.started) {
-                this.register();
                 this.processManager.success();
                 return;
             }
@@ -30,29 +27,7 @@ class Instance {
     }
 
     kill() {
-        this.deregister();
         this.child.kill();
-    }
-
-    register() {
-        let systemConfig = Config.load('system');
-        let instances = systemConfig.get('instances', {});
-
-        instances[process.cwd()] = {
-            mode: process.env.NODE_ENV,
-            url: this.config.get('url'),
-            port: this.config.get('server.port'),
-            process: this.config.get('process')
-        };
-
-        systemConfig.set('instances', instances).save();
-    }
-
-    deregister() {
-        let systemConfig = Config.load('system');
-        let instances = systemConfig.get('instances', {});
-        delete instances[process.cwd()];
-        systemConfig.set('instances', instances).save();
     }
 }
